# Conflicts:
#	library/build/intermediates/blame/res/release/multi/values-af.json
#	library/build/intermediates/blame/res/release/multi/values-am.json
#	library/build/intermediates/blame/res/release/multi/values-ar.json
#	library/build/intermediates/blame/res/release/multi/values-az.json
#	library/build/intermediates/blame/res/release/multi/values-be.json
#	library/build/intermediates/blame/res/release/multi/values-bg.json
#	library/build/intermediates/blame/res/release/multi/values-bn.json
#	library/build/intermediates/blame/res/release/multi/values-bs.json
#	library/build/intermediates/blame/res/release/multi/values-ca.json
#	library/build/intermediates/blame/res/release/multi/values-cs.json
#	library/build/intermediates/blame/res/release/multi/values-da.json
#	library/build/intermediates/blame/res/release/multi/values-de.json
#	library/build/intermediates/blame/res/release/multi/values-el.json
#	library/build/intermediates/blame/res/release/multi/values-en-rGB.json
#	library/build/intermediates/blame/res/release/multi/values-es-rUS.json
#	library/build/intermediates/blame/res/release/multi/values-es.json
#	library/build/intermediates/blame/res/release/multi/values-et.json
#	library/build/intermediates/blame/res/release/multi/values-eu.json
#	library/build/intermediates/blame/res/release/multi/values-fa.json
#	library/build/intermediates/blame/res/release/multi/values-fi.json
#	library/build/intermediates/blame/res/release/multi/values-fr-rCA.json
#	library/build/intermediates/blame/res/release/multi/values-fr.json
#	library/build/intermediates/blame/res/release/multi/values-gl.json
#	library/build/intermediates/blame/res/release/multi/values-gu.json
#	library/build/intermediates/blame/res/release/multi/values-hi.json
#	library/build/intermediates/blame/res/release/multi/values-hr.json
#	library/build/intermediates/blame/res/release/multi/values-hu.json
#	library/build/intermediates/blame/res/release/multi/values-hy.json
#	library/build/intermediates/blame/res/release/multi/values-in.json
#	library/build/intermediates/blame/res/release/multi/values-is.json
#	library/build/intermediates/blame/res/release/multi/values-it.json
#	library/build/intermediates/blame/res/release/multi/values-iw.json
#	library/build/intermediates/blame/res/release/multi/values-ja.json
#	library/build/intermediates/blame/res/release/multi/values-ka.json
#	library/build/intermediates/blame/res/release/multi/values-kk.json
#	library/build/intermediates/blame/res/release/multi/values-km.json
#	library/build/intermediates/blame/res/release/multi/values-kn.json
#	library/build/intermediates/blame/res/release/multi/values-ko.json
#	library/build/intermediates/blame/res/release/multi/values-ky.json
#	library/build/intermediates/blame/res/release/multi/values-lo.json
#	library/build/intermediates/blame/res/release/multi/values-lt.json
#	library/build/intermediates/blame/res/release/multi/values-lv.json
#	library/build/intermediates/blame/res/release/multi/values-mk.json
#	library/build/intermediates/blame/res/release/multi/values-ml.json
#	library/build/intermediates/blame/res/release/multi/values-mn.json
#	library/build/intermediates/blame/res/release/multi/values-mr.json
#	library/build/intermediates/blame/res/release/multi/values-ms.json
#	library/build/intermediates/blame/res/release/multi/values-my.json
#	library/build/intermediates/blame/res/release/multi/values-nb.json
#	library/build/intermediates/blame/res/release/multi/values-ne.json
#	library/build/intermediates/blame/res/release/multi/values-nl.json
#	library/build/intermediates/blame/res/release/multi/values-pa.json
#	library/build/intermediates/blame/res/release/multi/values-pl.json
#	library/build/intermediates/blame/res/release/multi/values-pt-rBR.json
#	library/build/intermediates/blame/res/release/multi/values-pt-rPT.json
#	library/build/intermediates/blame/res/release/multi/values-ro.json
#	library/build/intermediates/blame/res/release/multi/values-ru.json
#	library/build/intermediates/blame/res/release/multi/values-si.json
#	library/build/intermediates/blame/res/release/multi/values-sk.json
#	library/build/intermediates/blame/res/release/multi/values-sl.json
#	library/build/intermediates/blame/res/release/multi/values-sq.json
#	library/build/intermediates/blame/res/release/multi/values-sr.json
#	library/build/intermediates/blame/res/release/multi/values-sv.json
#	library/build/intermediates/blame/res/release/multi/values-sw.json
#	library/build/intermediates/blame/res/release/multi/values-ta.json
#	library/build/intermediates/blame/res/release/multi/values-te.json
#	library/build/intermediates/blame/res/release/multi/values-th.json
#	library/build/intermediates/blame/res/release/multi/values-tl.json
#	library/build/intermediates/blame/res/release/multi/values-tr.json
#	library/build/intermediates/blame/res/release/multi/values-uk.json
#	library/build/intermediates/blame/res/release/multi/values-ur.json
#	library/build/intermediates/blame/res/release/multi/values-uz.json
#	library/build/intermediates/blame/res/release/multi/values-vi.json
#	library/build/intermediates/blame/res/release/multi/values-zh-rCN.json
#	library/build/intermediates/blame/res/release/multi/values-zh-rHK.json
#	library/build/intermediates/blame/res/release/multi/values-zh-rTW.json
#	library/build/intermediates/blame/res/release/multi/values-zu.json
#	library/build/intermediates/blame/res/release/multi/values.json
#	library/build/intermediates/blame/res/release/single/color.json
#	library/build/intermediates/blame/res/release/single/drawable-hdpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable-ldpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable-mdpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable-tvdpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable-xhdpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable-xxhdpi-v4.json
#	library/build/intermediates/blame/res/release/single/drawable.json
#	library/build/intermediates/blame/res/release/single/layout.json
#	library/build/intermediates/blame/res/release/single/raw.json
#	library/build/intermediates/bundles/release/classes.jar
#	library/build/intermediates/incremental/mergeReleaseAssets/merger.xml
#	library/build/intermediates/incremental/mergeReleaseJniLibFolders/merger.xml
#	library/build/intermediates/incremental/mergeReleaseResources/merger.xml
#	library/build/intermediates/incremental/mergeReleaseShaders/merger.xml
#	library/build/intermediates/incremental/packageReleaseResources/merger.xml
#	library/build/outputs/aar/android-maps-utils-release.aar

+++ app/src/main/java/com/app/sample/messenger/model/Group.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Group implements Serializable {
-    private long id;
+    private int id;
     private String date;
     private String name;
     private String snippet;
@@ -13,7 +13,7 @@ public class Group implements Serializable {
     private int photo;
     private ArrayList<Friend> friends = new ArrayList<>();
 
-    public Group(long id, String date, String name, String snippet, String address, int photo) {
+    public Group(int id, String date, String name, String snippet, String address, int photo) {
         this.id = id;
         this.date = date;
         this.name = name;
@@ -22,7 +22,7 @@ public class Group implements Serializable {
         this.photo = photo;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
